[US_TN] deleting update_history from yamls for RECIDIVIZ_REFERENCE_staff_supervisor_and_caseload_roster (Recidiviz/recidiviz-data#35137)
## Description of the change

RECIDIVIZ_REFERENCE_staff_supervisor_and_caseload_roster is a manual
roster table fraught with many errors in data entry etc. However, in
prep of the TN insights launch, all historical versions of this data are
getting updated in order to support multiple supervisor for SCU
caseloads and SCU admin caseloads.

As such, old versions of these files that used to not have all the
columns....now _will_ have those columns since we will retroactively add
them in in order to support the new functionality.

therefore, this PR deleted the update_history flags in the yamls since
they will no longer be needed.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4596af73b97ee2a8a9b521b0ca36e807f03453f6