[calc dag] Move ingest_pipeline_should_run_in_dag out of KPO (Recidiviz/recidiviz-data#35120)
## Description of the change

Moving `ingest_pipeline_should_run_in_dag` out of kubernetes because of
various issues that's been causing the task to be flaky. As it stands
the function checks 1. is ingest launched in the state 2. are there
launchable views. Because we branch on states that have ingest launched
[here](https://github.com/Recidiviz/recidiviz-data/blob/main/recidiviz/airflow/dags/calculation_dag.py#L221)
the first condition will always be true so I've rebranded this check as
`check_region_has_launchable_ingest_views`

After this change is made we can remove the option to use xcom in KPO
and close out Recidiviz/recidiviz-data#31148

source visibility check is failing i'll fix it in a little bit

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33880

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a2a9915bacf7be8ea5acd3ac0a5c224b466be01e