Sort table references in direct_ingest_documentation_generator (Recidiviz/recidiviz-data#35167)
## Description of the change

Prep work for adding a smarter CI job to check if region documentation
needs to be regenerated.

Because table references are gathered dynamically when building region
config, there is no guaranteed ordering. Therefore a CI check to see if
the region documentation needs to be regenerated might fail even though
there are no actual differences between the files, just the ordering of
the table refs changed.

This PR is the output of making the sort change and running `pipenv run
python -m recidiviz.tools.docs.region_documentation_generator
--force-all`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#10343

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e3a64fd28004201d544ba432e496faf94207ad22