[US_AZ] Filter invalid conducting staff IDs from MH assessments view (Recidiviz/recidiviz-data#35256)
## Description of the change

There were 720 rows in the current results of this view that include the
`PERSON_ID` "13" as the conducting staff member. This is not a staff
member, but rather "Service Account", which implies the record was
administratively added or altered. Ingest is currently crashing because
there is no staff member associated with that ID; I don't want to ingest
this account as a staff member because it is not one.

Since this "13" doesn't actually tell us anything about who conducted
the assessment, this change NULLs out the `RECORD_USERID` for all rows
where it would otherwise be 13 in the source data to resolve the issue.

Tests:
- [x] run mappings script succeeds
- [x] query results to make sure that there are no other `RECORD_USERID`
values that do not appear in `normalized_state.state_staff_external_id`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c8a8619352f6bfbebd6f6e6d5daa4b465f0a0101