[Utils] Make some date utils explicitly timezone aware (Recidiviz/recidiviz-data#35074)
## Description of the change

Whenever we use `datetime.datetime.now()` or `date.today()` in the
context of our deployed environment, those values are returning times in
US/Eastern because that's the timezone configured on our Docker images.

This PR updates a few date helpers to explicitly call out that they're
returning current date in US/Eastern time, and also adds a new helper /
tests for getting the first day of the week (will use this in a
subsequent PR for aggregated metrics optimization).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#29291

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: da038f6d3567ac145d13cf4b4cd44ef3e0c9f0e2