[US_ME] Correct how calculate county jail start dates (Recidiviz/recidiviz-data#35275)
## Description of the change
The previous way we were selecting the number of detention days for
individuals who spent time in county jail was by selecting the largest
number of detention days. You'd think this is a positive choice, but
actually by selecting a large number of days, it made certain
individuals ineligible for SCCP because it appeared their sentences were
longer, and thus the required number of days to be served is higher.
This is ME specific -- if you are sentenced to 5 years or more, certain
opps require you to serve 2/3 of your sentence, but if it's 5 years or
less, then you only have to serve 1/2 the sentence.

In all, this change affects 21 residents, and shortens the time required
for everyone. Only a couple of people actually get additionally
opportunities for this, but still! You can see the full list of changes
[here](https://docs.google.com/spreadsheets/d/1pvx2qVw21saqStnOK-CCagSCOwU8znLLGJ6PKaE8cqQ/edit?gid=827248226#gid=827248226).

This does resolve the issue for the person mentioned in
https://app.zenhub.com/workspaces/all-pods-6324d1d66be1c200119ab65a/issues/gh/recidiviz/recidiviz-dashboards/6639.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Relates to recidiviz-dashboards/Recidiviz/recidiviz-data#6639
(https://app.zenhub.com/workspaces/all-pods-6324d1d66be1c200119ab65a/issues/gh/recidiviz/recidiviz-dashboards/6639)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d8326f8a30fe97b4abea8b3f7935c7f23872df69