[Vitals] exclude pretrial offices from vitals in ND (Recidiviz/recidiviz-data#35269)
## Description of the change

This change was asked about in the Insights x Nebula meeting today since
it sounds like ND's really wanting to see it, so I went ahead and did it
now directly in the old queries since it was pretty easy (too easy? am i
missing something?)

Loaded to sandbox at `dana_vitals_exclude_pretrial`. compare views
script shows 19 removals, all of which are an officer in a pretrial
office or a pretrial office itself, and 1 change for the statewide
metrics going up since they aren't taking the pretrial ones into account
anymore.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#34387

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 080877108f2c21288fe2e4519444bf83a1effd01