[Workflows] Convert opportunity config field objects to lists (Recidiviz/recidiviz-data#35215)
## Description of the change

In the workflows opportunity configs, converts fields that used to be
objects with meaningful orders into arrays. Though Python and JS both
preserve object key order, the systems in between don't always: for
instance, by default Flask sorts keys to improve cacheability. This PR
reenables this sorting behavior.

This PR does include a db migration with some hand-written sql to
convert the objects that already exist. I've tested this on my local dev
instance, and it works as expected. I don't know if it's possible to
create a test suite for a migration.

Recidiviz-dashboards is prepared for this change: it can accept either
objects or arrays.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33956

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 19454d82710560f80eca7eb0a00d77da944112e6