[Workflows] Opportunity Admin Panel Improvements (Recidiviz/recidiviz-data#35218)
## Description of the change
A variety of changes to the Workflows admin panel, which fall into three
main buckets:
- Opportunity configs now have two description fields: a Variant
Description and a Revision Description. The Variant Description can keep
track of what's special about this feature variant, while the Revision
Description can note what's been changed in a particular edit. For
existing configs, the current description field has been migrated to
Revision Description, and a Variant Description is autofilled from the
feature variant.
- Instead of a bunch of ugly JSON dumps, the config view now mirrors the
tag structure of the form. The layout for complex fields is still a work
in progress, but it looks better than it did. Eventually I'd like to
unify the view and form-- I'm still working out the best way to do it,
but this seems like a step in the right direction.
- A bunch of wording and layout changes to make things clearer. For
example, inactive configs are hidden by default, "Ineligible Criteria
Copy" is relabeled "Almost Eligible Criteria Copy", the button to make
changes to a config is now `Edit` rather than `Duplicate`, etc.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 990489d7696dde6b11a2f200ff0913abca3085f1