[Insights] add new location fields on supervisor product view (Recidiviz/recidiviz-data#35331)
## Description of the change

My end goal is to change ND Insights over to office instead of district,
since that's what we use in other ND products as people's locations.
However, we have constraints with CA where they want to use office for
the supervisor page, but district for the supervisor list page. This
solution seemed reasonable to me- instead of exporting "district" and
"unit" fields, just export whatever district we should use for the
supervisor page, and whatever district we should use for the list page.
Order of operations:
1. This PR
2. (after prod deploy) Update BE to add to schema and return new fields
3. (after following prod deploy) Update FE to use the new fields
4. (after FE prod deploy) Update BE + product view to stop returning old
fields

There might be a way to speed up ^ but we're not in a rush and this is
safest.

Loaded to sandbox at `dana_insights_locations`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#35330

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 354e3241295843fc67e94ab7cac4f3586c5bf8a1