[KPIs] Small refactor to _create_or_update_view_and_materialize_if_necessary (Recidiviz/recidiviz-data#35334)
## Description of the change

Makes a few preparatory changes for logging view-level results as part
of the view update:
* Introduces a new `BigQueryViewMaterializationResult` class that stores
not just the materialized table, but also the completed materialization
QueryJob so that we can surface info like slot_millis.
* Rebrand `force_materialize` to `materialize_changed_views_only`
everywhere (flipping the value of the flag). Add
`--materialize_changed_views_only` as a flag to `load_views_to_sandbox`
but otherwise defaults `materialize_changed_views_only=False` everywhere
(cc @justkunz )
* Refactor `_create_or_update_view_and_materialize_if_necessary`
slightly to simplify the amount of branching logic. A subsequent change
will return the new `materialization_result` along with the update
status so we can persist the info.

Tested that running `load_views_to_sandbox` with the same args a second
time rematerialized views, but running a third time with the new flag
skipped materialization.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#34767

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a4b6aad878e8eae0392789dc8a31d6cb68af1ce9